# documents/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "User API",
        "description": "API para gestionar usuarios en el sistema",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/member": {
            "post": {
                "tags": [
                    "Member"
                ],
                "summary": "Crear un socio",
                "operationId": "b6ab178ca3ef94de1c4494ecf5d9b46d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Sandra Montero"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "sandra1@gmail.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "1234567890"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Socio creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "string",
                                            "example": "Member created"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/member/{id}": {
            "get": {
                "tags": [
                    "Member"
                ],
                "summary": "Obtener detalles de un socio",
                "operationId": "e97c87b45b35d4078f3c757746acb6e9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del socio",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del socio",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "string",
                                            "example": "Member show"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": {
                                                "name": "John Doe",
                                                "email": "sandra1@gmail.com",
                                                "phone": "1234567890"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Socio no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Member"
                ],
                "summary": "Actualizar un socio",
                "operationId": "7880c1161d8da7951c657d3af2999295",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del socio",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Jane Montero"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "sandra@example.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "0987654321"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Socio actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "string",
                                            "example": "Member updated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error al actualizar el socio"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Member"
                ],
                "summary": "Eliminar un socio",
                "operationId": "2ade32a098a358d9a93c899d33ac851e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del socio",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Socio eliminado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "string",
                                            "example": "Member deleted"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}